Failed attempt to enforce methods on all routes using type checker
Tried to use extensible records, but it didn't work. Routes without methods still type-check.
For missing method, Elm is just inferring the type rather than raising an error.

Difficulty is that we want to use the URL functions like s and </> both before and after specifying the method.
For things like 'string' and 'int' we could deliberately generate the wrong type, to be fixed by 'm' later.
But not every route needs to have a URL parameter (e.g. articles list)